Fix console spam after crash
